+++ pkg/reconciler/ksvc/ksvc_suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package knservice
+package ksvc
 
 import (
 	knservingv1alpha1 "github.com/knative/serving/pkg/apis/serving/v1alpha1"
